requirements-docker.txt
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==42.0.6
cssselect==1.2.0
iniconfig==2.0.0
jmespath==1.0.1
lxml==5.2.1
packaging==24.0
parsel==1.9.1
pdfminer.six==20231228
pluggy==1.5.0
pycparser==2.22
pycryptodome==3.20.0
pytest==8.2.0
python-dateutil==2.9.0.post0
setuptools==69.5.1
six==1.16.0
w3lib==2.1.2
wheel==0.43.0
